refactor: раскидать Movies по компонентам
